Add DistributionInstaller subclass to MkosiArgs
This subclass is meant to be subclassed by individual distributions, to hook
into the image creation process.

Individual steps in the image creation can check whether they have an instance
of a DistributionInstaller or MkosiArgs and can change their behaviour this why
and either do, what they did so far or pass on to the hooks on the
DistributionInstaller.

In the long term the general case can be folded into DistributionInstaller, but
the present thin wrapper allows for a piecemeal split.
