+++ mkosi/__init__.py
@@ -111,6 +111,7 @@ from .backend import (
     workspace,
     write_grub_config,
 )
+from .distributions import DistributionInstaller, configure_dracut
 from .manifest import Manifest
 
 __version__ = "13"
@@ -132,21 +133,6 @@ PathString = Union[Path, str]
 MKOSI_COMMANDS_NEED_BUILD = (Verb.shell, Verb.boot, Verb.qemu, Verb.serve)
 MKOSI_COMMANDS_SUDO = (Verb.build, Verb.clean, Verb.shell, Verb.boot, Verb.qemu, Verb.serve)
 MKOSI_COMMANDS_CMDLINE = (Verb.build, Verb.shell, Verb.boot, Verb.qemu, Verb.ssh)
-
-DRACUT_SYSTEMD_EXTRAS = [
-    "/usr/bin/systemd-repart",
-    "/usr/lib/systemd/system-generators/systemd-veritysetup-generator",
-    "/usr/lib/systemd/system/initrd-root-fs.target.wants/systemd-repart.service",
-    "/usr/lib/systemd/system/initrd-usr-fs.target",
-    "/usr/lib/systemd/system/sysinit.target.wants/veritysetup.target",
-    "/usr/lib/systemd/system/systemd-repart.service",
-    "/usr/lib/systemd/system/systemd-volatile-root.service",
-    "/usr/lib/systemd/system/veritysetup.target",
-    "/usr/lib/systemd/systemd-veritysetup",
-    "/usr/lib/systemd/systemd-volatile-root",
-    "/usr/bin/systemd-ask-password",
-    "/usr/bin/systemd-tty-ask-password-agent"
-]
 
 
 def write_resource(
@@ -1593,46 +1579,6 @@ def mount_cache(args: MkosiArgs, root: Path) -> Iterator[None]:
 
 def umount(where: Path) -> None:
     run(["umount", "--recursive", "-n", where])
-
-
-def configure_dracut(args: MkosiArgs, packages: Set[str], root: Path) -> None:
-    if "dracut" not in packages:
-        return
-
-    dracut_dir = root / "etc/dracut.conf.d"
-    dracut_dir.mkdir(mode=0o755)
-
-    dracut_dir.joinpath('30-mkosi-hostonly.conf').write_text(
-        f'hostonly={yes_no(args.hostonly_initrd)}\n'
-        'hostonly_default_device=no\n'
-    )
-
-    dracut_dir.joinpath("30-mkosi-qemu.conf").write_text('add_dracutmodules+=" qemu "\n')
-
-    with dracut_dir.joinpath("30-mkosi-systemd-extras.conf").open("w") as f:
-        for extra in DRACUT_SYSTEMD_EXTRAS:
-            f.write(f'install_optional_items+=" {extra} "\n')
-
-    if args.hostonly_initrd:
-        dracut_dir.joinpath("30-mkosi-filesystem.conf").write_text(
-            f'filesystems+=" {(args.output_format.needed_kernel_module())} "\n'
-        )
-
-    if args.get_partition(PartitionIdentifier.esp):
-        # These distros need uefi_stub configured explicitly for dracut to find the systemd-boot uefi stub.
-        if args.distribution in (Distribution.ubuntu,
-                                 Distribution.debian,
-                                 Distribution.mageia,
-                                 Distribution.openmandriva,
-                                 Distribution.gentoo):
-            dracut_dir.joinpath("30-mkosi-uefi-stub.conf").write_text(
-                "uefi_stub=/usr/lib/systemd/boot/efi/linuxx64.efi.stub\n"
-            )
-
-        # efivarfs must be present in order to GPT root discovery work
-        dracut_dir.joinpath("30-mkosi-efivarfs.conf").write_text(
-            '[[ $(modinfo -k "$kernel" -F filename efivarfs 2>/dev/null) == /* ]] && add_drivers+=" efivarfs "\n'
-        )
 
 
 def prepare_tree_root(args: MkosiArgs, root: Path) -> None:
